Data.List.Maybe

Description

A module that works with lists of Maybes.

Documentation

type MaybeList a = [Maybe a]

A type synonym for a list of Maybe values.

concatMapMaybes

Arguments

:: Foldable f

=> (a -> MaybeList b)
The function that is used to perform a mapping from the items to a list of Maybe values.

-> f a
The given Foldable of items for which we perform a mapping.

-> [b]
As result a list of items that were wrapped in a Just data constructor.

Concatenate the items wrapped in a Just when we perform a mapping and concatenate the results of these mappings.